practice/15-CASE-创建你的RAG问答（LangChain）/pyproject.toml
[project]
name = "rag-qa-system"
version = "0.1.0"
description = "基于LangChain的RAG问答系统"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.1.0",
    "faiss-cpu>=1.7.4",
    "PyPDF2>=3.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]